refac: split da clients into separate mods
