+++ utils/host/src/witness_generation/mod.rs
@@ -1,3 +1,5 @@
+pub mod celestia_da_client;
 pub mod client;
+pub mod eth_da_client;
 pub mod online_blob_store;
 pub mod preimage_witness_collector;
